+++ command/agent/command.go
@@ -29,8 +29,8 @@ import (
 	flaghelper "github.com/hashicorp/nomad/helper/flag-helpers"
 	gatedwriter "github.com/hashicorp/nomad/helper/gated-writer"
 	"github.com/hashicorp/nomad/helper/logging"
+	"github.com/hashicorp/nomad/helper/winsvc"
 	"github.com/hashicorp/nomad/nomad/structs/config"
-	"github.com/hashicorp/nomad/service_os"
 	"github.com/hashicorp/nomad/version"
 	"github.com/mitchellh/cli"
 	"github.com/posener/complete"
@@ -778,7 +778,7 @@ WAIT:
 	select {
 	case s := <-signalCh:
 		sig = s
-	case <-service_os.Shutdown_Channel():
+	case <-winsvc.Shutdown_Channel():
 		sig = os.Interrupt
 	case <-c.ShutdownCh:
 		sig = os.Interrupt
